requirements.txt
transformers==4.30.2
torch==2.0.1
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
python-dotenv==1.0.0
pyyaml==6.0
tqdm==4.65.0
jinja2==3.1.2
pytest==7.3.1
